Added in metadata for wifi boards, also fixed a bug where the wifi state of the board was not correctly being unmarshalled
