+++ internal/boards/board_test.go
@@ -17,9 +17,11 @@ func TestCombine(t *testing.T) {
 	assert.Equal(t, "B7-400X", response.Boards[1].Name)
 	assert.Equal(t, "C1-100X", response.Boards[2].Name)
 	assert.Equal(t, "D4-200S", response.Boards[3].Name)
-	assert.Equal(t, "Low_Power", response.Boards[4].Name)
-	assert.Equal(t, 5, response.Metadata.Totals.Boards)
-	assert.Equal(t, 2, response.Metadata.Totals.Vendors)
+	assert.Equal(t, "Shrug Board", response.Boards[4].Name)
+	assert.Equal(t, "Low_Power", response.Boards[5].Name)
+	assert.Equal(t, 6, response.Metadata.Totals.Boards)
+	assert.Equal(t, 3, response.Metadata.Totals.Vendors)
+	assert.Equal(t, 3, response.Metadata.Totals.WifiEnabled)
 	assert.Equal(t, false, response.Metadata.Errors.HasErrors)
 }
 
@@ -32,7 +34,8 @@ func TestCombine_WithInvalidJsonFilesInFolders(t *testing.T) {
 	assert.Equal(t, 3, response.Metadata.Totals.Boards)
 	assert.Equal(t, 2, response.Metadata.Totals.Vendors)
 	assert.Equal(t, true, response.Metadata.Errors.HasErrors)
-	assert.Equal(t, 1, response.Metadata.Errors.FileReadErrors)
+	assert.Equal(t, 2, response.Metadata.Errors.FileReadErrors)
+	assert.Contains(t, response.Metadata.Errors.Files, "testdata/invalidMix/boards-broken.json")
 	assert.Contains(t, response.Metadata.Errors.Files, "testdata/invalidMix/invalid.json")
 }
