+++ internal/boards/testdata/invalidMix/boards-broken.json
@@ -1,0 +1,3 @@
+{
+  in
+}
